+++ lib/niku.dart
@@ -2,5 +2,6 @@ library niku;
 
 export 'widget/widget.dart';
 export 'objects/objects.dart';
+export 'proxy/proxy.dart';
 
 export 'extra/extra.dart';