+++ lib/proxy/proxy.dart
@@ -1,2 +1,2 @@
-export 'buttonStyle.dart';
-export 'textStyle.dart';
+export 'buttonProxy.dart';
+export 'textFormFieldProxy.dart';